--- packages/presentational-components/src/components/CommandPalette/index.ts
@@ -1,2 +1,0 @@
-export * from './CommandPalette';
-export * from './CommandButtonRow';
