+++ packages/presentational-components/src/components/command-palette/index.ts
@@ -1,0 +1,2 @@
+export * from "./command-palette";
+export * from "./command-button-row";
